% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maumauR-package.R
\docType{package}
\name{maumauR-package}
\alias{maumauR}
\alias{maumauR-package}
\title{maumauR: Simulation and Analysis of the DB Mau Mau Card Game}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a configurable engine, Monte Carlo simulator, and strategy framework for the "der kleine ICE" Mau Mau card game, a German children's variant of Mau Mau with draw-two, skip, and wild cards. Includes win-rate estimation by seat and player count, pluggable artificial strategies for comparison, game-length and draw-count distribution summaries, and an interactive Shiny application for playing against the simulated opponents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-heller/maumauR}
  \item \url{https://r-heller.github.io/maumauR/}
  \item Report bugs at \url{https://github.com/r-heller/maumauR/issues}
}

}
\author{
\strong{Maintainer}: Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

Authors:
\itemize{
  \item Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})
}

}
\keyword{internal}
